refactoring static library
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@46160 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ retex/renderer-ios/ReTeX/Classes/graphics/ColorI.m
@@ -41,8 +41,8 @@
     jint color = 0;
 
     for (int i = 0; i < CGColorGetNumberOfComponents(_color); i++) {
-        NSUInteger component = (NSUInteger) (components[i] * 255 + 0.5);
-        color = color << 8 | component;
+        int component = (int) (components[i] * 255 + 0.5);
+        color = (color << 8) | component;
     }
     return color;
 }
